backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ccxt>=4.3.0
sqlalchemy==2.0.36
aiosqlite==0.19.0
apscheduler==3.10.4
python-dotenv==1.0.0
pydantic-settings==2.1.0
websockets==12.0
httpx==0.25.2
cryptography==41.0.7
